+++ CHANGELOG.md
@@ -7,11 +7,19 @@ and this library adheres to Rust's notion of
 
 ## [Unreleased]
 
-## [0.1.0] - 2025-01-06
-Initial release.
+## [0.3.0] - 2025-06-03
+### Added
+- Support for the asset description hash in the issuance bundle
+- Structures for describing transfer and burn information, ability to convert them into transactions
+- Support for scenarios with arbitrary number of accounts
+- Additional scenario for a three party test case
 
 ## [0.2.0] - 2025-02-28
 ### Added
 - OrchardBundle enum
 - first_issuance flag
 - action-group-based serialization
+
+
+## [0.1.0] - 2025-01-06
+Initial release.